fix(scope): give up on an unstable digest
What we need to do is keep running the digest for some acceptable
number of iterations. If
the scope is still changing after those iterations we have to throw our
hands up and declare
it’s probably never going to stabilize. At that point we might as well
throw an exception,
since whatever the state of the scope is it’s unlikely to be what the
user intended.