perf(scope): short-circuit digest when the last watch is clean
What we can do to cut the number of executions in half is to keep
track of the last watch we have seen that was dirty. Then,
whenever we encounter a clean watch,we check whether it’s also
the last watch we have seen that was dirty. If so, it means a
full round has passed where no watch has been dirty. In that case
there is no need to proceed to the end of the current round. We
can exit immediately instead.