fix(scope): ues value-based dirty-checking
Value-based dirty-checking implies that if the old or new values are
objects or arrays we
have to iterate through everything contained in them. If there’s any
difference in the two
values, the watcher is dirty. If the value has other objects or arrays
nested within, those
will also be recursively compared by value.