feat(scope): add $watchGroup method for observing a set of expressions
The $watchGroup function takes several watch functions wrapped in an
array, and a single listener function. The idea is that when any of
the watch functions given in the array detects a change, the listener
function is invoked. The listener function is given the new and old
values of the watches wrapped in arrays, in the order of the original
watch functions.